🏗️ refactor: move layout to layouts
