+++ wini/src/lib.rs
@@ -2,7 +2,7 @@
 #![feature(impl_trait_in_fn_trait_return)]
 pub mod components;
 pub mod cron;
-pub mod layout;
+pub mod layouts;
 pub mod pages;
 pub mod server;
 pub mod shared;
